requirements.txt
pypdf
torch
transformers
sentence-transformers
llama-index
langchain-community
langchain-core
langchain-openai
streamlit
faiss-cpu
openai
